Historyfragment

+++ app/src/main/java/project/st991377867/marcin/data/model/ItemEntry.kt
@@ -5,13 +5,13 @@ import java.util.Date
 class ItemEntry(
     _date: Date,
     _itemName: String,
-    _itemQuantity: Int,
+    _itemQuantity: Double,
     _kCal: Int
 )
 {
     val date: Date
     val itemName: String
-    val itemQuantity: Int
+    val itemQuantity: Double
     val kCal: Int
 
     init {
